@@ -8,7 +8,7 @@ import Form from './components/Form';
 // ReactDOM.render(<Form />, document.getElementById('editor'));
 
 import { createRoot } from 'react-dom/client';
-const container = document.getElementById('editor');
+const container = document.getElementById('editor')!;
 const root = createRoot(container);
 root.render(
 	<StrictMode>
